SPARKC-507: Infinite Retries and  Mark all Idmepotent Queries
Changes our retry policy to retry all exceptions which can possible be
temporary. This will move burden to users to detect when the system is
beyond hope but will allow a momentarily troubled cluster recover from
issues.

Some queries will not be retried regardless of the retry policy. To fix
this we mark all of our idempotent queries so the driver will retry
them.